requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.3
Werkzeug==2.3.7
WTForms==3.0.1
qrcode==7.4.2
Pillow==10.0.1
python-dateutil==2.8.2
